requirements.txt
apache-airflow==2.5.2
black==23.3.0
python-dotenv==0.21.1
rdflib==6.3.2
requests==2.28.2